bug🐜: web gpu wasn't ready but a backend is required for creating a detector
web gpu is one of the possible backends in tfjs:
1. cpu
2. webgl
3. webgpu

4. other ones for node.js

this only happened in brave not edge or chrome. WHY?
- Brave updated to chromium 113 which supports WebGPU. 🦁🦁
- Chrome using chromium 112 🤡
- Edge using chromium 112 🤡
🤯🤯🤯🤯🤯

Why wasn't WebGPU ready in backend?
- WebGPU requires some time to be ready ... 🙄
https://www.npmjs.com/package/@tensorflow/tfjs-backend-webpu
- unlike WebGL 🤦
https://www.npmjs.com/package/@tensorflow/tfjs-backend-webgl

THE FIX! 🔨🛠️🪛👷‍♂️:
wait for the backend to be ready using built-in function.

alternatives:
could also set backend manually and wait for that too. E.g. tf.setBackend('webgpu')

or just set to webgl always as that doesn't have a wait time for some reason
tf.setBackend('webgl')