updating format of KRAs
the end key rep area will not be included in the KRA array if it is the same as the start.

so when showing key rep areas on video, only need to show the KRAs array 👍👍

and when editing exercise, I will show the inferred end KRA as well, so it can be edited simultaneously with the start KRA.

Also removing unecessary param for deleteKRA function in KRAsList.svelte.

and so when the start/end same setting is changed, the side effects are much simpler to implement (in FocusLimb.svelte)

finally, the limb rep counter has to be adapted slightly for this new format to work. the keyAreas property for limbRepCounter instances CONTAINS the inferred end KRA.

updating the validation during class constructor

SORRY for length of message 🥲😂🥲😢😭