issue: key rep areas will be in different locations
and have diff dimensions when scale (video size to client element size scale) is different.

NOTE: the rep counter doesn't need the scale to work correctly, scale is only required when showing KRA's is involved.

fix: scale KRA dimensions so KRAs are visually in the same place.

- added scaling/unscaling methods to KeyRepArea class

- simplified the format of key rep area's corners to:
absTopLeft and all other corners are relative to it.

- improved type safety for keypoint and KRAs.